requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
bokeh @ file:///C:/Windows/TEMP/abs_4a259bc2-ed05-4a1f-808e-ac712cc0900cddqp8sp7/croots/recipe/bokeh_1658136660686/work
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy==0.7.0
cachetools==5.2.0
category-encoders @ file:///C:/Windows/TEMP/abs_96hpd46mm1/croots/recipe/category_encoders_1659117695542/work
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer==2.1.1
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
commonmark==0.9.1
cryptography @ file:///C:/ci/cryptography_1652101729823/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
daal4py==2021.4.0
dask @ file:///C:/ci/dask-core_1658497112560/work
dask-searchcv==0.2.0
decorator==5.1.1
distributed @ file:///C:/ci/distributed_1658523963030/work
distro @ file:///tmp/build/80754af9/distro_1610479853857/work
entrypoints==0.4
fonttools==4.25.0
fsspec @ file:///C:/Windows/TEMP/abs_a2mhnomvfy/croots/recipe/fsspec_1659972224540/work
future @ file:///C:/ci_310/future_1642081234309/work
gitdb==4.0.9
GitPython==3.1.27
HeapDict @ file:///Users/ktietz/demo/mc3/conda-bld/heapdict_1630598515714/work
idna==3.4
imagecodecs @ file:///C:/ci_310/imagecodecs_1642082516950/work
imageio @ file:///C:/Windows/TEMP/abs_24c1b783-7540-4ca9-a1b1-0e8aa8e6ae64hb79ssux/croots/recipe/imageio_1658785038775/work
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==4.16.0
kiwisolver @ file:///C:/ci/kiwisolver_1653292359073/work
locket @ file:///C:/ci/locket_1652904090946/work
lz4 @ file:///C:/ci_310/lz4_1643300078932/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib @ file:///C:/ci/matplotlib-suite_1660169687702/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
msgpack @ file:///C:/ci/msgpack-python_1652348582618/work
munkres==1.1.4
networkx @ file:///C:/ci/networkx_1657716953747/work
numexpr @ file:///C:/Windows/Temp/abs_e2036a32-9fe9-47f3-a04c-dbb1c232ba4b334exiur/croots/recipe/numexpr_1656940304835/work
numpy==1.23.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.4
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
patsy==0.5.2
Pillow==9.2.0
protobuf==3.20.2
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pyarrow==9.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
pyrsistent==0.18.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets @ file:///C:/ci/pywavelets_1648728096239/work
PyYAML==6.0
requests==2.28.1
rich==12.5.1
scikit-build @ file:///C:/ci_310/scikit-build_1642112544589/work
scikit-image @ file:///C:/ci/scikit-image_1648214343198/work
scikit-learn==1.1.2
scikit-learn-intelex==2021.20220113.184028
scikit-rf @ file:///tmp/build/80754af9/scikit-rf_1624298598240/work
scipy==1.9.1
semver==2.13.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
statsmodels==0.13.2
streamlit==1.12.2
tblib @ file:///Users/ktietz/demo/mc3/conda-bld/tblib_1629402031467/work
threadpoolctl==3.1.0
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
zict==2.1.0
zipp==3.8.1
